Preserve Rust-pump executor settlement (#297)
Keep the worker-owned writer descriptor out of completion cleanup, so a
reused file-descriptor slot cannot be closed after settlement.

Move terminal span completion and cancellation waiting into their focused
modules, preserving the executor boundary while keeping the pipeline module
within its size limit. Wrap the benchmark workflow conditions to satisfy the
repository YAML lint gate.
